move module/seats -> store/index.js
